requirements.txt
# Semantic Document Search
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# AI & Vector Database
sentence-transformers>=2.2.0
chromadb>=0.4.22

# Document Processing
pymupdf>=1.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Frontend
streamlit>=1.30.0
requests>=2.31.0
